backend/pyproject.toml
[project]
name = "novaops"
version = "0.1.0"
description = "Multi-Agent DevOps Command Center powered by Amazon Nova"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "strands-agents>=0.3.0",
    "strands-agents-tools>=0.2.0",
    "boto3>=1.35.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "websockets>=14.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
    "mangum>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
